Worked on to fixed not showing contacts permission dialog box in contacts screen,Worked on to close keyboard when users click on submit button in all calculator screen,Worked on to change int value to double value in all calculator screen,Worked on to change ui in retirement calculator screen,Worked on to fix keyboard issue in add member screen

+++ lib/blocs/emisipcalculator/emisip_calculator_event.dart
@@ -4,18 +4,17 @@ part of 'emisip_calculator_bloc.dart';
 abstract class EMISIPCalculatorEvent {}
 
 class EMISIPData extends EMISIPCalculatorEvent {
-
   final String name;
   final int loanAmount;
   final int noOfYear;
-  final int loanInterestRate;
-  final int interestRateOnInvestment;
+  final double loanInterestRate;
+  final double interestRateOnInvestment;
 
-  EMISIPData(
-      {required this.name,
-        required this.loanAmount,
-        required this.noOfYear,
-        required this.loanInterestRate,
-        required this.interestRateOnInvestment,
-      });
+  EMISIPData({
+    required this.name,
+    required this.loanAmount,
+    required this.noOfYear,
+    required this.loanInterestRate,
+    required this.interestRateOnInvestment,
+  });
 }
